wip refactoring for generic faker processign implementation
