+++ src/treecipe/src/CollectionsApiService/ICollectionsApiJsonStructure.ts
@@ -1,0 +1,10 @@
+// interface CollectionsApiStructure {
+//     allOrNone: boolean;
+//     records: any[];
+// }
+
+interface CollectionsApiJsonStructure {
+    allOrNone: boolean;
+    records: any[];
+}
+
