+++ src/treecipe/src/FakerRecipeProcessor/IFakerRecipeProcessor.ts
@@ -1,0 +1,10 @@
+
+export interface IFakerRecipeProcessor {
+
+    isRecipeProcessorSetup(): Promise<boolean>;
+    generateFakeDataBySelectedRecipeFile(fullRecipeFileNamePath: string): Promise<string>;
+    transformFakerJsonDataToCollectionApiFormattedFilesBySObject(fakerContent: string): Map<string, CollectionsApiJsonStructure>;
+
+    
+}
+
